fix: hide guest header CTAs for authed users and add returnUrl
- Extract GuestHeader into a 'use client' component so it can read
  auth state via useAuth and current path via usePathname
- Hide Create Account / Log In links when isAuthenticated (fixes
  authenticated non-member users seeing irrelevant CTAs on 403 path)
- Append returnUrl to both auth links, matching GuestPromoBanner and
  PrivateChannelLockedPane behavior
- Wrap server name in min-w-0 flex-1 + truncate to prevent overflow
  when CTA buttons are present
